Backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
Werkzeug==2.3.4
Pillow==10.0.1
pillow-avif-plugin==1.3.1
supabase==2.0.2
python-dotenv==1.0.0
    